requirements.txt
feedparser==6.0.11
requests>=2.32.0
PyYAML>=6.0
jinja2>=3.1.2
python-dateutil>=2.9.0
markdown-it-py>=3.0.0
